+++ dffml/df/types.py
@@ -129,6 +129,7 @@ class Operation(NamedTuple, Entrypoint):
     expand: Optional[List[str]] = []
     instance_name: Optional[str] = None
     validator: bool = False
+    retry: int = 0
 
     def export(self):
         exported = {
@@ -138,6 +139,7 @@ class Operation(NamedTuple, Entrypoint):
             "conditions": self.conditions.copy(),
             "stage": self.stage.value,
             "expand": self.expand.copy(),
+            "retry": self.retry,
         }
         for to_string in ["inputs", "outputs"]:
             exported[to_string] = dict(